v0.3 — Systematic bug hunt: 11 bugs fixed + test suite
Audit findings fixed (from 3 parallel Explore agents):
- CRITICAL: MP3Source._launch_current self-deadlock (Lock→metadata()→same Lock)
- CRITICAL: YouTubeSource.close() ValueError on double-close of stdout pipe
- CRITICAL: MP3Source.read_chunk infinite loop when all files corrupt
- CRITICAL: YouTubeSource.open() leaks yt-dlp process if ffmpeg fails
- HIGH: _run_tui/_run_headless no try/finally — subprocesses leaked on exception
- HIGH: Player._refill_loop no try/except — any processor error kills audio silently
- MEDIUM: scipy.sosfilt float64 promotion — all modes now return float32 via .astype()
- MEDIUM: FMMode blend + AMMode mono np.mean float64 creep — dtype=np.float32
- MEDIUM: NaN guard added to pipeline clip + player int16 cast
- LOW: YouTubeSource unused self._lock removal (dead code)
- LOW: _current_metadata field dead code removed (metadata() computes inline, not cached)

Test suite (tests/):
- conftest.py — shared helpers: tone generators, WAV writer, audio assertions
- test_smoke.py — 6 tests: all modes process tone, silence, stereo, noise monotonic,
  NaN after 60s, near-0dBFS clip. All pass.
- test_numerical.py — 7 tests: DC offset, gain staging, noise monotonic,
  de-emphasis cutoff, AM bandwidth, fading mean, noise generators. All pass.

13/13 tests passing. Zero crashes from edge-case inputs.

+++ src/modes/am.py
@@ -39,14 +39,14 @@ class AMMode(RadioMode):
         audio = self._mix_static(audio, signal_db)
         audio = self._inject_crackle(audio, signal_db)
         audio = self._apply_fading(audio, signal_db)
-        return audio
+        return audio.astype(np.float32, copy=False)
 
     # ---- processing stages ----
 
     def _apply_mono(self, audio: np.ndarray) -> np.ndarray:
         """Mix stereo to dual-mono (average L+R)."""
         if audio.shape[1] >= 2:
-            mono = np.mean(audio[:, :2], axis=1, keepdims=True)
+            mono = np.mean(audio[:, :2], axis=1, keepdims=True, dtype=np.float32)
             return np.repeat(mono, audio.shape[1], axis=1)
         return audio
