+++ src/modes/amhd.py
@@ -40,10 +40,10 @@ class AMHDMode(RadioMode):
 
         # Crossfade
         if blend <= 0.0:
-            return digital
+            return digital.astype(np.float32, copy=False)
         elif blend >= 1.0:
-            return analog
-        return (1.0 - blend) * digital + blend * analog
+            return analog.astype(np.float32, copy=False)
+        return ((1.0 - blend) * digital + blend * analog).astype(np.float32, copy=False)
 
     # ---- paths ----
